[GraphEditor] Remove enabling/disabling pins depending on edge's visibility
When edges could exist while being invisible in the graph, pins needed
to be enabled/disabled manually according to the visibility of the edge.
This was especially important when closing the application and removing
all the (hidden) edges as the pins needed to be properly destryed as well.

Edges are not hidden anymore, and any attribute that should be hidden
but is in fact connected remains visible as long as the connection exists.