[GraphEditor] AttributePin: Do not start dragging edge on `pressed`
Prior to this commit, the edge that was being dragged to be connected
appeared as soon as the `pressed` signal was emitted by the mouse. Now
that the user can actually click (press & release) without dragging the
mouse to expand or collapse groups, we want to be able to distinguish
cases where the user presses the mouse to click (to expand/collapse
groups) from those where the user presses the mouse to drag it (to
connect edges).

Instead of triggering the drag as soon as the `pressed` signal is
emitted, we wait to see if the mouse moves significantly enough to
indicate the will to drag; an arbitrary limit of 10 pixels along the
X- or Y-axis is used to determine that the user means to drag the mouse.